ujuzi/requirements.txt
asgiref==3.8.1
Authlib==1.3.2
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-crontab==0.7.1
django-google-sso==7.1.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
flake8==7.1.1
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
isort==5.13.2
loguru==0.7.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
